chore: build lib
