+++ lib/index.bak.d.ts
@@ -1,0 +1,2 @@
+import 'dotenv/config';
+//# sourceMappingURL=index.bak.d.ts.map
