+++ lib/index.d.ts
@@ -1,2 +1,4 @@
-import 'dotenv/config';
+export * from './bot';
+export * from './nlp';
+export * from './models';
 //# sourceMappingURL=index.d.ts.map
