+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "qute-nlp",
-  "version": "0.9.0",
+  "version": "0.9.1",
   "description": "QUTE NLP engine",
   "main": "lib/index.js",
   "scripts": {
